Sharing very preliminary Arjun's selenium scraper
This needs to be finished and documented. But it may serve as inspiration for folks working toward a universal scraper. 
Some notes on this: The complexity of current web design means that traditional methods of web scraping (e.g., wget) don’t often capture text in a functional way. Selenium--on which this script is based--goes a long way to address these issues. It uses a “headless” or automated browser (usually Chrome) to fully interpret web pages so they can be accurately scraped and parsed. 
It doesn’t depend on HTML tags (e.g., “</p>” tags to denote paragraphs) to interpret structure, which is an advantage given that many web designers don’t use these consistently, well, or at all.